Progress

+++ 1sat/main.go
@@ -36,7 +36,7 @@ type Msg struct {
 	Height      uint32
 	Hash        string
 	Status      uint32
-	Idx         uint32
+	Idx         uint64
 	Transaction []byte
 }
 
@@ -127,7 +127,7 @@ func subscribe() {
 				msgQueue <- &Msg{
 					Id:          txResp.Id,
 					Height:      txResp.BlockHeight,
-					Idx:         uint32(txResp.BlockIndex),
+					Idx:         txResp.BlockIndex,
 					Transaction: txResp.Transaction,
 				}
 			},
